Add new project with core
