--- Configuration/BotConfig.cs
@@ -1,7 +1,0 @@
-﻿namespace Codewars_Bot.Configuration
-{
-    public class BotConfig
-    {
-        public string BotApiToken { get; set; }
-    }
-}
