--- Configuration/CodewarsConfig.cs
@@ -1,7 +1,0 @@
-﻿namespace Codewars_Bot.Configuration
-{
-    public class CodewarsConfig
-    {
-        public string CodewarsApiToken { get; set; }
-    }
-}
