--- Configuration/DbConfig.cs
@@ -1,7 +1,0 @@
-﻿namespace Codewars_Bot.Configuration
-{
-    public class DbConfig
-    {
-        public string DbConnectionString { get; set; }
-    }
-}
