--- Configuration/Environment.cs
@@ -1,14 +1,0 @@
-﻿namespace Codewars_Bot.Configuration
-{
-    public static class Environment
-    {
-        public static string Name()
-        {
-#if DEBUG
-            return "Debug";
-#else
-            return "Release";
-#endif
-        }
-    }
-}
