--- Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Codewars_Bot.WebApiApplication" Language="C#" %>
