+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Configuration/BotConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace ITKPI.CodewarsBot.Api.Configuration
+{
+    public class BotConfig
+    {
+        public string BotApiToken { get; set; }
+    }
+}
