+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Configuration/CodewarsConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace ITKPI.CodewarsBot.Api.Configuration
+{
+    public class CodewarsConfig
+    {
+        public string CodewarsApiToken { get; set; }
+    }
+}
