+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Configuration/DbConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace ITKPI.CodewarsBot.Api.Configuration
+{
+    public class DbConfig
+    {
+        public string DbConnectionString { get; set; }
+    }
+}
