+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Configuration/Environment.cs
@@ -1,0 +1,14 @@
+﻿namespace ITKPI.CodewarsBot.Api.Configuration
+{
+    public static class Environment
+    {
+        public static string Name()
+        {
+#if DEBUG
+            return "Debug";
+#else
+            return "Release";
+#endif
+        }
+    }
+}
