+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Contracts/ICodewarsService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using ITKPI.CodewarsBot.Api.Models;
+
+namespace ITKPI.CodewarsBot.Api.Contracts
+{
+	public interface ICodewarsService
+	{
+		Task<CodewarsResponseModel> GetCodewarsUser(string username);
+	}
+}
