+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Contracts/IDatabaseService.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+using ITKPI.CodewarsBot.Api.Models;
+
+namespace ITKPI.CodewarsBot.Api.Contracts
+{
+	public interface IDatabaseService
+	{
+		void AuditMessageInDatabase(string message);
+		List<string> GetWeeklyRating(bool onlyActiveUsers);
+		List<string> GetTotalRating();
+		List<string> GetWeeklyPoints(int userId);
+		string DeleteUserInfo(int userId);
+		string SaveUserToDatabase(UserModel user);
+		UserModel GetUserById(int userId);
+	}
+}
