+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Models/AuditMessageModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace ITKPI.CodewarsBot.Api.Models
+{
+    public class AuditMessageModel
+    {
+        public string Message { get; set; }
+    }
+}
