+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Models/CodewarsResponseModel.cs
@@ -1,0 +1,8 @@
+﻿namespace ITKPI.CodewarsBot.Api.Models
+{
+    public class CodewarsResponseModel
+    {
+        public string Name { get; set; }
+        public int Honor { get; set; }
+    }
+}
