+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Models/WeekModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ITKPI.CodewarsBot.Api.Models
+{
+    public class WeekModel
+    {
+        public int Id { get; set; }
+        public int WeekNumber { get; set; }
+        public DateTime EndDate { get; set; }
+    }
+}
