+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Models/WeeklyPointsModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ITKPI.CodewarsBot.Api.Models
+{
+	public class WeeklyPointsModel
+	{
+		public int WeekNumber { get; set; }
+		public DateTime EndDate { get; set; }
+		public int Points { get; set; }
+	}
+}
