+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/Models/WeeklyRatingUserModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace ITKPI.CodewarsBot.Api.Models
+{
+    public class WeeklyRatingUserModel
+    {
+        public int Id { get; set; }
+        public string CodewarsUsername { get; set; }
+        public int WeekNumber { get; set; }
+        public int Points { get; set; }
+    }
+}
