+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/appsettings.Local.json
@@ -1,0 +1,6 @@
+{
+  "MigrationDBName": "Codewars_Local",
+  "RunMigration": true,
+  "DBConnectionString": "Server=localhost,9123;User Id=sa;Password='SomeLongPasswordForRequ1rements!';",
+  "MigrationFilePath": "../../../DatabaseMigration.sql"
+}
