+++ ITKPI.CodewarsBot.Api/ITKPI.CodewarsBot.Api/appsettings.json
@@ -1,0 +1,3 @@
+﻿{
+
+}
