--- Models/AuditMessageModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Codewars_Bot.Models
-{
-    public class AuditMessageModel
-    {
-        public string Message { get; set; }
-    }
-}
