--- Models/CodewarsResponseModel.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Codewars_Bot.Models
-{
-    public class CodewarsResponseModel
-    {
-        public string Name { get; set; }
-        public int Honor { get; set; }
-    }
-}
