--- Models/WeekModel.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Codewars_Bot.Models
-{
-    public class WeekModel
-    {
-        public int Id { get; set; }
-        public int WeekNumber { get; set; }
-        public DateTime EndDate { get; set; }
-    }
-}
