--- appsettings.Debug.json
@@ -1,5 +1,0 @@
-﻿{
-  "RunMigration": true,
-  "DBConnectionString": "Server=localhost,9123;User Id=sa;Password='SomeLongPasswordForRequ1rements!';",
-  "MigrationFilePath": "DatabaseMigration.sql"
-}
