--- appsettings.Release.json
@@ -1,3 +1,0 @@
-{
-
-}
