--- appsettings.json
@@ -1,4 +1,0 @@
-﻿
-{
-
-}
