--- packages.config
@@ -1,48 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Autofac" version="4.8.1" targetFramework="net472" />
-  <package id="Autofac.WebApi2" version="4.2.0" targetFramework="net472" />
-  <package id="Chronic.Signed" version="0.3.2" targetFramework="net472" />
-  <package id="Dapper" version="1.50.5" targetFramework="net472" />
-  <package id="IdentityModel" version="3.10.5" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.Bot.Builder" version="3.19.1.44133" targetFramework="net472" />
-  <package id="Microsoft.Bot.Connector" version="3.19.1.44133" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.EnvironmentVariables" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.FileExtensions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Json" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileProviders.Abstractions" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileProviders.Physical" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileSystemGlobbing" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging" version="2.1.1" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="2.1.1" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Options" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Primitives" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="4.4.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.JsonWebTokens" version="5.3.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Logging" version="5.3.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Protocol.Extensions" version="1.0.4.403061554" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Protocols" version="5.3.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Protocols.OpenIdConnect" version="5.3.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Tokens" version="5.3.0" targetFramework="net472" />
-  <package id="Microsoft.Rest.ClientRuntime" version="2.3.18" targetFramework="net472" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.3" targetFramework="net472" />
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net472" />
-  <package id="SourceLink.Create.CommandLine" version="2.8.1" targetFramework="net472" developmentDependency="true" />
-  <package id="System.Buffers" version="4.5.0" targetFramework="net472" />
-  <package id="System.ComponentModel.Annotations" version="4.5.0" targetFramework="net472" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="5.3.0" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.1" targetFramework="net472" />
-  <package id="System.Net.Requests" version="4.3.0" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net472" />
-  <package id="System.Text.Encodings.Web" version="4.5.0" targetFramework="net472" />
-  <package id="Telegram.Bot" version="14.10.0" targetFramework="net472" />
-</packages>
